### 2024-11-22 #7515 @warrickball
Note also that the paper is a bit long. [Our docs say](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain)

> The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper.

The "Overview of SpecpolFlow" in particular can probably be pared down somewhat.
