### 2024-12-06 9652aae
Draft module for the Bz fit
