### 2024-12-06 8efda52
Update Highlights.md
Updating citation for HD 54879 paper
