### 2025-02-01 65b0890
extended the .fits file converter for ESPaDOnS to work with i.fits files
