+++ specpolFlow/converters/espadons.py
@@ -8,13 +8,105 @@ import numpy as np
 import astropy.io.fits as fits
 from .. import obsSpec as spf
 
-def espadons(flist, flistout=None, writeSpecHeader=False):
+def espadons(flist, flistout=None, ftype=None, writeSpecHeader=False):
     """
-    Convert a list of .fits files in the CADC ESPaDOnS format into
-    text .s files. 
+    Convert a list of .fits files in the CADC ESPaDOnS format
+    into text .s files.  This function can process both intensity only i.fits
+    spectra and intensity+polarization p.fits spectra.
 
-    The code provides two files in .s format:  
-     
+    By default the type of file will be inferred based on the name, if it ends
+    with either i.fits or p.fits.  Alternatively, the ftype function parameter
+    can be set to either 'i' or 'p' for those respective file types.
+    
+    This function generates two files in a .s format:  
+    
+    * The UPENA normalized spectrum (n.s), with automated radial velocity corrections from the telluric lines.
+    * The UPENA unnormalized spectrum (u.s), using the automated radial velocity correction from the normalized spectrum.
+    This is done starting from the unnormalized spectrum without the automated 
+    radial velocity correction, to which we apply the radial velocity
+    correction determined from the normalized spectrum. The reason behind this
+    is that the UPENA automated radial velocity determination performed on 
+    unnormalized spectra does not produce consistently reliable results. 
+    The content of the fits header is also saved in a .out text file.
+    If flistout=None (default), the files are written at the same path as
+    the fits-format data, with the '.fits' stripped, 
+    and 'n.s' and 'u.s' appended to the filename root. 
+    If flistout is a list of paths ending with a rootname, the files will be
+    saved at that path with that rootname, and 'n.s' and 'u.s' appended to
+    the filename root.
+
+    :param flist: (list of strings) a list of ESPaDOnS filenames
+    :param flistout: (list of strings) optional, list of output file rootnames.
+                     If this is not given output file names will be based on
+                     the input names.
+    :param ftype: optional, set the type of .fits file to process.  
+                  By default, if this is not set, the type will be inferred
+                  from the file name, if the name ends in i.fits or p.fits. 
+                  This can be set to the text string 'i' for intensity
+                  (no polarization) spectra, or to 'p' for spectropolarimetric
+                  (intensity + polarization) spectra.
+    :param writeSpecHeader: Flag for whether to write two lines of header in
+                            the .s text file (True = include header)
+    :return: list of normalized Spectrum objects,
+             and a list of unnormalized Spectrum objects
+    """
+
+    if isinstance(flist, str): flist = [flist,]
+    #Some basic error checking
+    if not isinstance(flist, list):
+        raise ValueError('in espadons(), the flist argument must be a Python '
+                         'list of input file names (or a single file name)')
+    if not (flistout is None):
+        if isinstance(flistout, str): flistout = [flistout,]
+        if not isinstance(flistout, list):
+            raise ValueError('in espadons(), the flistout argument must be a '
+                             'Python list of output file names')
+        if len(flist) != len(flistout):
+            raise ValueError('in espadons(), the list of input file names and '
+                             'list of output file names must have the same length')
+
+    speclist_n = []
+    speclist_u = []
+    for i, fname in enumerate(flist):
+        #If an output file name exists use it
+        if flistout is None:
+            fnameOut = None
+        else:
+            fnameOut = flistout[i]
+            
+        #If no file type is given, try guessing from the file name
+        _ftype = ftype
[…]
