+++ specpolFlow/__init__.py
@@ -5,5 +5,5 @@ from .lineList import *
 from .cleanMaskUI import cleanMaskUI
 from . import converters
 
-__version__ = "0.4.3"
+__version__ = "0.4.4"
 
