### 2025-02-10 8c38fab
Development
- Created a notebook for the description of the available models.
- Arranged the dipole_model class
- Wrote the notes on the dipole_model in the notebook.
(Next step will be the clean up the cos_model and write the doc for it)
