### 2025-02-11 #7515 @folsomcp
Has there been any progress on finding an editor for this paper?  It has been a couple months now.  @warrickball do you happen to have any updates?  If there is anything we can do to help speed things up please let us know.

In the mean time we will try to edit the paper down some more.
