### 2025-02-11 bec9dfd
edited the paper for length, although it may still be too long
