+++ paper/paper.md
@@ -63,19 +63,19 @@ bibliography: paper.bib
 
 # Summary
 
-Spectropolarimetry, the observation of polarization and intensity as a function of wavelength, is a powerful tool in stellar astrophysics. It is particularly useful for characterizing the distribution of circumstellar material and for tracing the influence of magnetic fields on a host star and its environment. Maintaining modern, flexible, and accessible computational tools that enable spectropolarimetric studies is thus essential. The `SpecpolFlow` package is a new, completely Pythonic workflow for analyzing stellar spectropolarimetric observations. Its suite of tools provides a user-friendly interface for working with data from an assortment of instruments and telescopes. `SpecpolFlow` contains tools for spectral normalization and visualization, the extraction of Least-Squares Deconvolution (LSD) profiles, the generation and optimization of line masks for LSD analyses, and the calculation of longitudinal magnetic field measurements from the LSD profiles. It also provides Python classes for the manipulation of spectropolarimetric products. The `SpecpolFlow` website includes an array of tutorials that guide users through common analytic analysis using the software. `SpecpolFlow` is distributed as a free, open-source package, with fully documented tools (via an API and command line interface) which are actively maintained by a team of contributors. 
+Spectropolarimetry, the observation of polarization and intensity as a function of wavelength, is a powerful tool in stellar astrophysics. It is particularly useful for characterizing stars and circumstellar material, and for tracing the influence of magnetic fields on a host star and its environment. Maintaining modern, flexible, and accessible computational tools that enable spectropolarimetric studies is thus essential. The `SpecpolFlow` package is a new, completely Pythonic workflow for analyzing stellar spectropolarimetric observations. Its suite of tools provides a user-friendly interface for working with data from an assortment of instruments and telescopes. `SpecpolFlow` contains tools for spectral normalization and visualization, the extraction of Least-Squares Deconvolution (LSD) profiles, the generation and optimization of line masks for LSD analyses, and the calculation of longitudinal magnetic field measurements from the LSD profiles. It also provides Python classes for the manipulation of spectropolarimetric products. The `SpecpolFlow` website includes an array of tutorials that guide users through common analytic analysis using the software. `SpecpolFlow` is distributed as a free, open-source package, with fully documented tools (via an API and command line interface) which are actively maintained by a team of contributors. 
 
 # Statement of need
 
-Spectropolarimetry is an essential observational technique in stellar astrophysics, which is used to study the surface properties of stars and their environments. Symmetry-breaking phenomena like stellar magnetic fields leave an imprint on polarized spectra.  Magnetic fields are present in most classes of stars throughout their evolution [@DonatiLandstreet2009; @Mestel2012]. Cool stars (with convective envelopes) have a theoretical magnetic incidence of 100%, and display a wide range of observed field strengths, driven by variations in their internal dynamos [@Reiners2012]. Nearly 10% of hot stars (with radiative envelopes) also harbour strong magnetic fields [@Grunhut2017; @Sikora2019]. Furthermore, magnetic fields are found in evolved giants and compact stellar remnants such as white dwarfs and pulsars [@Ferrario2015]. Spectropolarimetry is a valuable tool for characterizing the strength, orientation, and topology of these fields. Therefore the maintenance and dissemination of computational tools enabling this technique are a key element of research.
[…]
