### 2025-02-11 4da8eda
Finished the docs for the cos model
