### 2025-02-11 5289f8d
Started on the harmonics model class
