### 2025-02-12 #7515 @warrickball
I'm afraid so @folsomcp. If no-one else step forwards, I should have another space within the next week or so. If you can recommend any reviewers, either from our [reviewer database](https://reviewers.joss.theoj.org/), previous JOSS publications (see the authors of the articles suggested [above](https://github.com/openjournals/joss-reviews/issues/7515#issuecomment-2494657868), though these aren't guaranteed to be useful...) or your own network, that'll help whichever editor picks this up get started with looking for reviewers when the time comes.

Sorry for the wait!
