### 2025-02-20 #7515 @folsomcp
Thanks for the update!

For potential reviewers, here are some colleagues who have experience with similar analysis techniques to the ones used in the paper, and who might be willing to provide a report.

Julien Morin (Université de Montpellier)
Pascal Petit (Institut de Recherche en Astrophysique et Planétologie, Toulouse)
Coralie Neiner (Paris Observatory)
Sandra Jeffers (Max Planck Institute for Solar System Research)
Alexandre David-Uraz (Central Michigan University)
Oleg Kochukhov (Uppsala University)
Stephen Marsden (University of Southern Queensland)

At a quick look those people don't seem to be in the reviewer database.

In the papers above, the paper by gully, reviewed by bmorris3 and wtgee seems to be the most similar to our project.  So perhaps one of them could be suitable.
