### 2025-02-21 39fda0c
Update Highlights.md
Updating publications
