### 2025-02-21 e74c915
updated calc_Bz
commented out redundant variable
