### 2025-02-21 308997d
Implemented the harmonics fit
