### 2025-02-24 #7515 @folsomcp
Thanks for being the editor on for this!

I should have read that policy before suggesting reviewers.  oops!

On my list  Julien Morin, Pascal Petit, Sandra Jeffers, and Stephen Marsden are BCool members.  (It is an un-funded collaboration, but it leads to some publications together).   In a similar vein, Oleg Kochukhov is part of the EXOHOST collaboration that I am also involved with (although we're not directly working together on a project at the moment).  Coralie Neiner and Alexandre David-Uraz  are not members of those collaborations, although I have published with them.

If you want to avoid anyone I have been a coauthor with in the last 4 years, that narrows down the list of people I can recommend quite a bit.  Maybe: James Sikora or Viktor Khalack could be possibilities?

I would rather avoid a reviewer from the Potsdam stellar physics group.  I was recently involved in a competing paper that contradicted some of their work, and so I'm not 100% sure they would be unbiased.

One limitation with my suggestions is that these are people with astronomy expertise (particularly spectropolarimetric observations of stars or stellar magnetic fields) but not necessarily software engineering expertise.
