### 2025-02-26 #7515 @folsomcp
One of our coauthors suggested Heloise Stevance as another possibility.  She is not really involved in the stellar magnetism community, but has more experience with software development, and familiarity with spectropolarimetry.  She seems to be in the reviewer database as HeloiseS.
