### 2025-03-02 c40f433
fixed a bug in the SPIRou converter, where it would fail when encountering an order containing entirely nan
