### 2025-03-08 5daa269
fixed a bug in the SPIRou converter, when the first order of the spectrum is entirely nan
