### 2025-03-11 #7891 @editorialbot
[REVIEW]: SpecpolFlow: a new software package for spectropolarimetry using Python
**Submitting author:** @folsomcp (<a href="http://orcid.org/0000-0002-9023-7890">Colin Folsom</a>)
**Repository:** https://github.com/folsomcp/specpolFlow
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
