### 2025-03-11 #7891 @warrickball
@folsomcp, I noticed that `cloc` thinks there's a lot of Assembly in your code. I had a closer look and this all seems to be stuff in the `docs-jb` folder, so I suspect there's a bunch of files there that are generated for the Jupyter Book (?) but don't need to be tracked in the repo. If possible, it'd be good to remove them—they're only likely to confuse—but it's not strictly a JOSS requirement. I'm not familiar with building Jupyter Books so I accept that you might need to track all those files.
