### 2025-03-12 674402a
added a doppler_shift() funciton to the Spectrum class
