### 2025-03-12 3896080
added a function to convolve a Spectrum with a Gaussinan instrumental profile
