### 2025-03-13 41f5e7f
made some optimizations to the Spectrum.convolveR() function
