### 2025-03-17 #7891 @folsomcp
@DerekAstro and @gully, thanks for agreeing to review our paper!

@warrickball I think `cloc` is misidentifying some files as Assembly.  There are some example input spectrum files in the `docs-jb` folder, which are used in the tutorials contained in the Jupyter Notebooks.  The convention people have used from previous spectropolarimetric data analysis software is to use a `.s` suffix for these files (although they are basically just a `.dat` file).  An unintended side effect of this is that some automatic file type identification tools think those files are assembly.  The same issue seems to be present in Github's 'Languages' breakdown.

Up to now this hasn't been an issue, we haven't been using any tools that rely on automatically identifying those file types.  But if it seems like too much potential confusion we could rename the files.  Or we could possibly move the example input files to a different repository.  (That would be less convenient for just running the Jupyter Notebooks, but would help the awkwardness that the current repository is mostly example input files by file size).
