### 2025-03-17 #7891 @warrickball
Ah, I see what you mean. I take my criticism back completely. I'd assumed they were build artefacts for the Jupyter Book but on closer inspection the docs folder is nice and clean, comprising only source. You don't need to do anything, and thanks for the explanation.
