### 2025-03-28 folsomcp/specpolFlow#3 @gully
JOSS Review March-April 2025
Hello 👋

This Issue tracks ongoing comments as part of the JOSS review located here:
https://github.com/openjournals/joss-reviews/issues/7891

The editorial guidance from JOSS encourages these discussion to here in the target repo, rather than the JOSS-hosted review thread.   The review guidance also encourages posting comments as they arise and not necessarily all in one fell-swoop.  I will plan to post multiple GitHub Issues to isolate the discussions thematically, and I may use this thread to aggregate the various topics in a checklist style.

1. Adoption of GitHub Issues or other Issue tracking system
The GitHub Issues (namely the system currently used for this report) appears to lack previous engagement.  The adoption of GitHub Issues is not necessarily a requirement of JOSS, nor a requisite to meet the broad standards of open source software.  Rather, the adoption of such systems generally conveys a forum for community engagement and serves as a destination to learn about upcoming or planned features, or known limitations etc.  If such planned features or known bugs/limitations exist, it would appear prudent to record them in some format, GitHub Issues or otherwise.
