### 2025-03-30 folsomcp/specpolFlow#3 @folsomcp
Hi, thanks for helping to review out paper for JOSS.  Please do let us know what issues you run into, or where there is room for improvement!  GitHub's issue tracking is the most convenient place for that, even if it may take a few days for me to get back to you.
