### 2025-04-18 160ead5
added calc_ew, calc_V_R, and _plot_ew functions
