+++ specpolFlow/profileLSD.py
@@ -1088,6 +1088,271 @@ class LSD:
             return True
         else:
             return False
+
+    def calc_ew(self, cog='I', norm='auto', lambda0=None,
+                velrange=None, ewwidth=None, fullOutput=True, plot=True):
+        '''
+        Helper function to return the equivalent width of Stokes I for
+        this LSD profile, for a given continuum level
+
+        :param cog: The value, or calculation method for the center of gravity.
+                    The choices are:
+                    'I': center of gravity of I,
+                    'V': center of gravity of V,
+                    'IV': center of gravity of I*V,
+                    'min': velocity of the minimum of I,
+                    or a float: a user defined value in km/s.
+        :param norm: calculation method for the continuum. The choices are:
+                    'auto': the median of I outside of velrange (if defined)
+                    or the full range (if velrange is not defined),
+                    or float: a user defined value to use for Ic.
+        :param lambda0: wavelength of the line center in nanometers (default=500).
+                    For an LSD profile, this is the lambda value the LSD
+                    profile shape was scaled with.
+        :param velrange: range of velocity to use for the determination of the
+                    line center and the continuum. If not defined, the whole
+                    range will be used. 
+        :param ewwidth: distance from the line center for the EW calculation.
+                One element = same on each side of line center.
+                Two elements, left and right of line center.
+                Not defined: using velrange.
+        :param fullOutput: If True, Return also the error of the equivalent
+                            width
+        :param plot: If True, Return also a figure
+                        
+        :return: the equivalent width in nanometers
+        '''
+        # Velrange is used to identify the position of the line,
+        # for calculating the cog, and for calculating the position
+        # of the continuum.
+        # If Velrange is not defined, it will use the whole range.
+        # The range for calculating EW itself is controlled by ewwidth below
+        if velrange != None:
+            inside = np.logical_and(self.vel>=velrange[0], self.vel<=velrange[1])
+            lsd_in = self[inside]
+            lsd_out = self[np.logical_not(inside)]
+            # saving the range for plotting later.
+            p_ewwidth = np.copy(velrange)
+            p_velrange = np.copy(velrange)
+        else:
+            lsd_in=copy.deepcopy(self)
+            p_ewwidth = [self.vel.min(), self.vel.max()]
+            p_velrange = [self.vel.min(), self.vel.max()]
+
+        # Check if norm is a string.
+        if isinstance(norm, str):
+            print('using AUTO method for the normalization')
+            if velrange != None:
+                print('  using the median of the continuum outside of the line')
+                norm_val = np.median(lsd_out.specI)
+            else:
+                print('  no range in velocity given, using the median of the whole specI to determine continuum')
+                norm_val = np.median(lsd_in.specI)
+        else:
+            norm_val = copy.copy(norm)
+            print('using given norm value')
+
+        # Check if cog is a string.
+        if isinstance(cog, str):
+            # calculate the cog for the chosen method
+            if cog == 'I':
+                cog_val = lsd_in.cog_I(norm_val)
+            elif cog == 'min':
+                cog_val = lsd_in.cog_min()
+            elif cog == 'IV':
+                cog_val = lsd_in.cog_IV(norm_val)
[…]
