### 2025-04-25 folsomcp/specpolFlow#4 @gully
"Developer guide" in documentation, or local development installation
The Installation docs are great, thank you! One apparent hiccup is the absence of guidance for candidate developers, in other words if folks wish to contribute to the code through local development.  The pip installation route would not clone the repo, for example.

One way to rectify this would be a "developer docs" option.  For example, here is the developer guidance for Lightkurve:
https://lightkurve.github.io/lightkurve/development/index.html

I think the relevant features for the developer docs for this repo could include:
- How to install an editable version of the code (`git clone`, etc.)
- How to build the documentation locally
- Guidance on dependencies for local development

This is not a requriement of the JOSS review, just a "nice to have"
