### 2025-04-25 folsomcp/specpolFlow#4 @gully
The guidance from the JOSS review is:

> Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

I'd recommend adding these points somewhere in the Docs if they are not already
