### 2025-04-25 #7891 @gully
OK, I have largely completed my review, bravo! I was able to run all of the tutorials, except for one, which was a bonus sonification feature, and I do not consider that to be a show-stopper.  I have opened some Issues in the specpolFlow repo, non of which are required to be solved for this review, with the possible exception of "developer guidelines".  Even that is soft, since it's natural to mos developers to hit `pip install -e .` in a GitHub cloned repo.  But adding that guidance somewhere on the docs is nice.

I definitely applaud that the tutorials worked out of the box, without additional steps.  And the figures rendered great.  This sort of open source transparency for spectropolarimetry is super useful to the community.  Bravo!
