### 2025-04-25 e7532e8
Update Highlights.md
Updated paper citations
