+++ docs-jb/About/Highlights.md
@@ -24,5 +24,7 @@ Check out the links below to see papers that have used `SpecpolFlow`! If you've 
 1. [Magnetic field, magnetospheric accretion and candidate planet of the young star GM Aurigae observed with SPIRou.](https://ui.adsabs.harvard.edu/abs/2024MNRAS.533.2893Z/abstract) Zaire et al. (2024) MNRAS, 533.3, 2893. 
 2. [First Observation of the Complete Rotation Period of the Ultra-Slowly Rotating Magnetic O Star HD 54879.](https://ui.adsabs.harvard.edu/abs/2024arXiv240109722E/abstract) Erba, Folsom, et al. (2024) ApJ, 977, 84.
 3. [A multi-kiloGauss magnetic field driving the magnetospheric accretion process in EX Lupi.](https://ui.adsabs.harvard.edu/abs/2024arXiv240903322P/abstract) Pouilly et al. (2024) A&A, 691, A18.
-4. [Magnetic Fields or Overstable Convective Modes in HR 7495: Exploring the Underlying Causes of the Spike in the 'Hump & Spike' Features.](https://ui.adsabs.harvard.edu/abs/2025arXiv250211879A/abstract) Antoci et al. (2025) A&A, accepted. 
+4. [Magnetic Fields or Overstable Convective Modes in HR 7495: Exploring the Underlying Causes of the Spike in the 'Hump & Spike' Features.](https://ui.adsabs.harvard.edu/abs/2025arXiv250211879A/abstract) Antoci et al. (2025) A&A, 696, A111.
+5. [Dissimilar magnetically driven accretion on the components of V4046 Sagittarii](https://ui.adsabs.harvard.edu/abs/2025arXiv250410217P/abstract) Pouilly & Audard (2025) A&A, accepted. 
+ 
  
