### 2025-05-01 folsomcp/specpolFlow#4 @folsomcp
Thanks for the suggestion!  A "developer docs" or an "information for developers" page could be quite useful.  That certainly seems like a more helpful starting point than saying "just email us".

Also, thanks for the example from Lightkurve, they do have a pretty good version of that.

We will get working on that, and I'll update this thread when we have something.
