### 2025-05-02 06f44e8
added an Optional additional packages section to the Installation page of the docs
