### 2025-05-07 0bdf3b6
adding a rough draft guide for people interested contributing to the project (still somewhat incomplete)
