+++ docs-jb/Tutorials/Contributing.md
@@ -1,0 +1,72 @@
+# Contributing
+
+SpecpolFlow is open source software, developed by a community of volunteers.  And we could use your help!  
+
+There are several ways you can contribute to the project:
+* Report bugs, or potential errors -- We try to test the things carefully, but in any sufficiently complicated code mistakes can slip through, so let us know what problems you run into.
+* Suggestions -- If there is a feature that you would like to have, or a way to make SpecpolFlow easier to use, let us know.  We can't promise to implement every suggestion (we have limited time!) but we'd at least like to know what would be useful.
+* Contribute code -- If there is a feature you have, this is a good way to make sure it happens!  
+* Add documentation -- A software package is only as good as its documentation.  We try to go beyond the bare minimum, and show different ways to use SpecpolFlow, for people with different levels of experience.  So there is always room to improve or expand the documentation and tutorials.
+
+## Reporting issues and making suggestions
+
+If you report a problem, or if you just want to make a suggestion, an easy way to do that is through the GitHub [issue tracker](https://github.com/folsomcp/specpolFlow/issues).  You can create a new thread (with the 'New issue' button), and also see what issues people have been talking about.
+
+You can also contact us directly at specpolflow@gmail.com
+
+## Contributing code
+
+If you want to get serious about contributing code, then the a good place to start is cloning the GitHub repository.  However, if you're not familiar with Git, and just want to contribute a small bit of code or a stand alone function, you can post it in the GitHub [Issues page](https://github.com/folsomcp/specpolFlow/issues), or email us at specpolflow@gmail.com
+
+Inside the Git repository, the documentation is all in the `docs-jb` folder, with folders for the `About`, `GetStarted`, and `Tutorials` sections of the website.  The documentation is all in the form of Markdown files, and Jupyter notebooks.  The Python code for the project is all in the `specpolFlow` folder.
+
+### Setting up with Git
+
+To make a copy of SpecpolFlow that you can modify, the typical way to do that is to start by [forking](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/working-with-forks/about-forks) the repository.  This basically creates your own copy of the code on GitHub.  You can do that on the [GitHub page](https://github.com/folsomcp/specpolFlow) with the `fork` button at the top right.  
+
+You can then make a local copy of your version of SpecpolFlow using the git clone command, for example:
+```
+git clone https://github.com/YOUR-GITHUB-USERNAME/specpolFlow.git
+```
+That will download the code and documentation files, which you can look at and modify.
+
+If you want to install your local copy of SpecpolFlow, it is a good idea to first make a [virtual environment](https://docs.python.org/3/tutorial/venv.html) to install the code in, using either Python's venv or conda.  For example create a virtual envornment called `spf-dev` with venv:
+```
+python -m venv spf-dev-env
+```
+That will create a new folder `spf-dev-env`.  Then activate the virtual environment by running the activate script:
+```
+source spf-dev-env/bin/activate
+```
+This keeps your development version separate from the regular version of SpecpolFlow.  You can switch versions just by activating or deactivating (just type `deactivate`) this virtual environment.  If you are using an IDE like VSCode, it may be easier to set up this virtual environment from within the that program.  
+
+You can make an ['editable' install](https://setuptools.pypa.io/en/latest/userguide/development_mode.html) of your copy of SpecpolFlow with pip.  First go to the folder that has your copy of SpecpolFlow (it should have pyproject.toml and README.md), then run pip with the `-e` flag:
+```
+cd specpolFlow
+pip install -e .
+```
+(The . at the end tells pip to install the version 'here' in that specpolFlow folder, not from their servers.)
+
[…]
