### 2025-05-07 1821554
cleaned up potentially messy bit in a unused string in the LSD.calc_bis() function
