### 2025-05-08 a69aefc
added a few more details about installing pandas in the docs
