### 2025-05-08 c8dfcce
slightly tweaked formatting in a couple docstrings in converters/espadons.py
