+++ specpolFlow/converters/espadons.py
@@ -22,6 +22,7 @@ def espadons(flist, flistout=None, ftype=None, writeSpecHeader=False):
     
     * The UPENA normalized spectrum (n.s), with automated radial velocity corrections from the telluric lines.
     * The UPENA unnormalized spectrum (u.s), using the automated radial velocity correction from the normalized spectrum.
+    
     This is done starting from the unnormalized spectrum without the automated 
     radial velocity correction, to which we apply the radial velocity
     correction determined from the normalized spectrum. The reason behind this
@@ -109,6 +110,7 @@ def espadons_p(flist, flistout=None, writeSpecHeader=False):
     
     * The UPENA normalized spectrum (n.s), with automated radial velocity corrections from the telluric lines.
     * The UPENA unnormalized spectrum (u.s), using the automated radial velocity correction from the normalized spectrum.
+    
     This is done starting from the unnormalized spectrum without the automated 
     radial velocity correction, to which we apply the radial velocity
     correction determined from the normalized spectrum. The reason behind this
@@ -244,6 +246,7 @@ def espadons_i(flist, flistout=None, writeSpecHeader=False):
     
     * The UPENA normalized spectrum (n.s), with automated radial velocity corrections from the telluric lines.
     * The UPENA unnormalized spectrum (u.s), using the automated radial velocity correction from the normalized spectrum.
+    
     This is done starting from the unnormalized spectrum without the automated 
     radial velocity correction, to which we apply the radial velocity
     correction determined from the normalized spectrum. The reason behind this
