### 2025-05-08 fdc6bb1
first draft of contributing docs
