### 2025-05-08 #7891 @warrickball
Hi! I'm finally checking back in after a long Easter break.

@DerekAstro, have you had a chance to have a look at the paper and try out the code? Let me know if you have any questions about the process.

@gully & @folsomcp, I see various discussions in the code repo though it wasn't quickly clear to me what the outstanding actions are. If you could revisit @gully's [issues](https://github.com/folsomcp/specpolFlow/issues) we might be able to check a few more things off the list.
