### 2025-05-09 #7891 @folsomcp
First off, thanks @gully for the review, and thanks for the helpful comments!

We've made some progress over the last few days on the issues @gully raised, and I've added some discussion in the [issues](https://github.com/folsomcp/specpolFlow/issues) page.  I think we have solutions implemented for all the issues except for a webpage with developer guidelines.  I've got a rough draft of that, but it could use an editing pass and possibly a bit more feedback from team members before it goes live.   I hope to get that up in the next few days, then perhaps @gully can comment on if that provides a good solution, and if there are any outstanding issues remaining.
