### 2025-05-09 #7891 @DerekAstro
I have looked at the paper but haven't tried out the code yet. Apologies, but I was unexpectedly swamped. It'll be my priority next week tho.

Get Outlook for Android<https://aka.ms/AAb9ysg>
