+++ docs-jb/Tutorials/Contributing.md
@@ -1,70 +1,99 @@
-# Contributing
+# Contributing and reporting issues
 
-SpecpolFlow is open source software, developed by a community of volunteers.  And we could use your help!  
+SpecpolFlow is open source software, developed by a community of users.  And we could use your help!  
 
 There are several ways you can contribute to the project:
-* Report bugs, or potential errors -- We try to test the things carefully, but in any sufficiently complicated code mistakes can slip through, so let us know what problems you run into.
+* Report bugs, or potential errors -- We try to test things carefully, but mistakes can slip through, so let us know what problems you run into.
 * Suggestions -- If there is a feature that you would like to have, or a way to make SpecpolFlow easier to use, let us know.  We can't promise to implement every suggestion (we have limited time!) but we'd at least like to know what would be useful.
-* Contribute code -- If there is a feature you have, this is a good way to make sure it happens!  
-* Add documentation -- A software package is only as good as its documentation.  We try to go beyond the bare minimum, and show different ways to use SpecpolFlow, for people with different levels of experience.  So there is always room to improve or expand the documentation and tutorials.
+* Contribute code -- If there is a feature you'd like to have, this is a good way to make sure it happens!  If you find a solution to a bug, please share it with the community.
+* Add documentation -- A software package is only as good as its documentation.  We try to go beyond the bare minimum, with tutorials on using SpecpolFlow for people with different levels of experience.  There is always room to improve or expand the documentation and tutorials.
+
+If you would like to contribute to SpecpolFlow on a regular basis, contact us about joining the core development team.
 
 ## Reporting issues and making suggestions
 
-If you report a problem, or if you just want to make a suggestion, an easy way to do that is through the GitHub [issue tracker](https://github.com/folsomcp/specpolFlow/issues).  You can create a new thread (with the 'New issue' button), and also see what issues people have been talking about.
+If you want to report a problem, or just make a suggestion, an easy way to do that is through the GitHub [issue tracker](https://github.com/folsomcp/specpolFlow/issues).  You can create a new thread (with the 'New issue' button), and also see what issues people have been talking about.
 
 You can also contact us directly at specpolflow@gmail.com
 
+## Contributing documentation
+
+The documentation for SpecpolFlow is all in the form of Jupyter notebooks and Markdown files.  These are all in the [`docs-jb`](https://github.com/folsomcp/specpolFlow/tree/main/docs-jb) folder on GitHub.  Inside that are folders for the `About`, `GetStarted`, and `Tutorials` sections of the website.
+
+If you want to make smaller changes, you can just download a copy of the relevant file, edit it, and contact us through email or on the [Issues page](https://github.com/folsomcp/specpolFlow/issues).  If you want to make bigger changes across multiple files, or are just comfortable using Git, you can follow the directions in the 'Contributing code' section below to download all the code and documentation files.
+
+:::{note}
+The .md and Jupyter files contain some non-standard Markdown notation, using extensions from [MyST Markdown](https://myst-parser.readthedocs.io/).  This is used for rendering some features on the website with [Jupyter Book](https://jupyterbook.org/).  (Like this [note](https://jupyterbook.org/en/stable/content/content-blocks.html#notes-warnings-and-other-admonitions)!)
+:::
+
 ## Contributing code
 
-If you want to get serious about contributing code, then the a good place to start is cloning the GitHub repository.  However, if you're not familiar with Git, and just want to contribute a small bit of code or a stand alone function, you can post it in the GitHub [Issues page](https://github.com/folsomcp/specpolFlow/issues), or email us at specpolflow@gmail.com
[…]
