### 2025-05-14 f6bc5d2
Update Highlights.md
Updated bibliography; restructured to reverse chronological order
