### 2025-05-14 d5f0475
Update Highlights.md
Trying to fix list numbering...
