### 2025-05-15 9323a94
Fix a wrong variable name in Spectrum.doppler_shift function
