+++ docs-jb/_toc.yml
@@ -30,6 +30,9 @@ parts:
     - file: Tutorials/7-IndividualLine_Tutorial.ipynb
     - file: Tutorials/8-CalculateBisector_Tutorial.ipynb
     - file: Tutorials/9-LSDsonification_Tutorial.ipynb
+  - caption: Contributing
+    chapters:
+    - file: Contributing/Contributing.md
 
   - caption: API
     chapters:
