### 2025-05-17 folsomcp/specpolFlow#4 @folsomcp
We've added a "Contributing" section to the website, with some guidelines for reporting bugs, getting in touch with us, or contributing code.   ( 1588652b64b5b4b09961117ef06696e4a2579ee0 )

There is some fairly minimal advice on forking and cloning the repository, installing an editable copy of the package, and a bit of using git.  Then there are links for people who would like more complete explanations.

That's now live on the website: https://folsomcp.github.io/specpolFlow/Contributing/Contributing.html
But if there is anything obviously missing, let us know!
