### 2025-05-17 #7891 @folsomcp
Just to update this thread, we have added a '[Contributing and reporting issues](https://folsomcp.github.io/specpolFlow/Contributing/Contributing.html)' section to the website, which I think will resolve the last of @gully's points.
(I've updated the relevant [Issues thread](https://github.com/folsomcp/specpolFlow/issues/4) too.)
