+++ paper/paper.bib
@@ -281,7 +281,7 @@ archivePrefix = {arXiv},
          year = 2004,
        series = {Astrophysics and Space Science Library},
        volume = {307},
-          doi = {10.1007/978-1-4020-2415-3},
+          doi = {10.1007/1-4020-2415-0},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2004ASSL..307.....L},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
