### 2025-05-23 #7891 @DerekAstro
@warrickball I finally got to spend a bunch of time on this and have made good progress, so I anticipate being done over the weekend. After a few hiccups at the start (entirely due to inexperience on my part), things have been going very smoothly.
