### 2025-05-27 78ef263
Update index.md
Update citation instructions
