### 2025-05-27 0db902a
Update Highlights.md
Update citation instructions
