### 2025-06-05 #7891 @warrickball
> We have been thinking about adding some very simple unit testing to the code. There might be a way to also make that available to users to test their install?

Since you have various CLI tools already, you could perhaps add another CLI tool like `spf-test` or similar, that could handle running some sort of check. I think it's okay for that to involve downloading data (which the script can handle) in order to keep the Python package small. I haven't reviewed the package myself but I agree that it'd be good to disentangle issues gathering data from issues in the code itself.
