### 2025-06-05 #7891 @DerekAstro
> Thanks for completing the review [@DerekAstro](https://github.com/DerekAstro), and thanks for the positive comments!
>
> The sonification tutorial does seem to be causing problems pretty often. Mostly because installing `astronify` is kind of tricky and unreliable. It's a fun 'extra', but if it consistently causes trouble then maybe the thing to do is just remove that tutorial from the website, and not officially support interoperability with `astronify`.

That does seem like the simplest and quickest solution.

> We have been thinking about adding some very simple unit testing to the code. There might be a way to also make that available to users to test their install? Even if it involved a user downloading some input files and a testing Python script, that might still be useful.

I think that would be very helpful, even if it is pretty minimal. At least it would provide a new user (particularly a non-expert one) reassurance that the install had worked as intended.
