### 2025-06-16 290936b
fix a typo in the paper
