### 2025-06-16 8101566
removed the astronify tutorial, and references to astronify, untill installing that package is easier
