### 2025-06-27 c382adb
reorganizing branches for version 1.1 and harmonic fitting
