+++ harmonicFit/BzFit_module.py
@@ -1,415 +1,0 @@
-'''
-Module with tools for doing MCMC fitting of Bz data with MCMC
-'''
-
-import numpy as np
-import scipy
-import emcee
-
-class data:
-    '''
-    A helper class to hold a set of values (and associated error bars) as a function of JD. 
-    '''
-    
-    def __init__(self, jd, val, val_err):
-        self.jd = jd
-        self.val = val
-        self.val_err = val_err
-    
-    def __getitem__(self, key):
-        """
-        Returns an data object with only the values at the specified index(s).
-
-        :param key: the index or slice being checked
-
-        :rtype: data
-        """
-        jd_s = self.jd[key]
-        val_s = self.val[key]
-        val_err_s = self.val_err[key]
-        
-        slice = data(jd_s, val_s, val_err_s)
-        return slice
-
-    def __setitem__(self, key, newval):
-        """
-        Sets all values of the bzdata object at the specified location equal to the
-        input object's values.
-
-        :param key: the index or slice being overwritten
-        :param newval: LSD object used to replace the overwritten values
-        """
-        if not(isinstance(newval, data)):
-            raise TypeError()
-        else:
-            self.jd[key] = newval.jd
-            self.val[key] = newval.val
-            self.val_err[key] = newval.val_err
-
-    def foldrot(self, P, jd0):
-        '''
-        Returns the Rotation.Phase that corresponds to self.jd,
-        for the specified period and JD0.
-
-        Note that Rotation can be negative, if JD0 is after the JD. 
-        '''
-        return (self.jd - jd0)/P
-    
-    def foldphase(self, P, jd0):
-        '''
-        Return the phase that corresponds to the self.jd,
-        for the specified period and JD0. 
-        '''
-        phi = self.foldrot(P, jd0)
-        phi = phi - np.fix(phi)
-        phi[phi<0] = 1+phi[phi<0]
-
-        return phi
-
-class dipole_model:
-    '''
-    Helper function to define a Bpole model for Bz
-    '''
-
-    def __init__(self, i, beta, Bpole, u):
-        self.i=i
-        self.beta=beta
-        self.Bpole=Bpole
-        self.u=u
-
-        const = self.Bpole/20*(15+u)/(3-u)
-        A = const*np.sin(self.i)*np.sin(self.beta)
-        mean = const*np.cos(self.i)*np.cos(self.beta)
-        self.cos_model = cos_model(A, mean, 0.0)
-
-    def to_array(self):
-        '''
-        Helper function to return an array of parameters
-        to use with minimizing functions
-        '''
-        return [self.i, self.beta, self.Bpole, self.u]
-    
-    def to_dict(self):
-        d = {
-            'i':self.i,
-            'beta':self.beta,
-            'Bpole':self.Bpole,
-            'u':self.u,
-        }
-        return d
-
-    def get_model(self, phi):
-        '''
-        Function to return the Bz value for a given phi for this dipolar model
-
-        :param phi: The rotational phase in radiants
-        '''
-        Bz = self.cos_model.get_model(phi)
-        return Bz
-
-    def get_cosalpha(self, phi):
-        '''
-        Function to return cos(alpha), where alpha is the angle between the magnetic axis and the line of sight.
-        cos(alpha) = cos(beta)cos(i) + sin(beta)sin(i)cos(phi)
-
[…]
