+++ harmonicFit/Colin_periodSearch-big.py
@@ -1,387 +1,0 @@
-#!/usr/bin/python3
-#
-# Fit sinusoids to data, to generate a periodogram.
-
-import numpy as np
-import matplotlib.pyplot as plt
-
-def periodogram(times, vals, errs, order=1,
-                pStart=0.1, pEnd=100., pNum=10000, jd0=None,
-                save=False, plot=False):
-    """
-    Calculate a periodogram by fitting sinusoids to data.
-
-    This generates a periodogram by fitting sinusoids and their harmonics
-    (from calculated from a Fourier series) to observed data, for a
-    range of periods.  The result a periodogram in reduced chi^2 and period.
-
-    :param times: an array of times of the observation, typically
-                  barycentric Julian Dates
-    :param vals: an array of the observed values (e.g. Bz measurements)
-    :param errs: an array of the error bars for the observed values
-    :param order: the order of the Fourier series used to fit the data
-                  (1 = fundamental only, 2 = fundamental + first harmonic)
-    :param pStart: the starting period of the grid searched
-    :param pEnd: the ending period of the grid searched
-    :param pNum: the number of periods in the grid (logarithmically spaced)
-    :param jd0: the reference Julian Date (time) for rotation cycle 0.0.
-                If this is None the first observation is used.
-    :param save: a Boolean flag to save the periodogram to a file (periods.dat)
-    :param plot: a Boolean flag to display a plot of the periodogram
-
-    :return: an array containing the periods used, and an array containing
-             the reduced chi^2 for those periods.
-    """
-    #Check data for consistency
-    if times.size != vals.size or times.size != errs.size:
-        print('ERROR: difference in sizes if input data arrays: '
-              +'times {:} vals {:} errs {:}'.format(
-                  times.size, vals.size, errs.size))
-        return np.zeros(0), np.zeros(0)
-    if times.size <= 2*order+1:
-        print('ERROR: too few observations for the number of free parameters'
-              +' {:}, {:}\n'.format(times.size, 2*order+1))
-    elif times.size <= 2*order+3:
-        print('Warning: few observations for to the number of free parameters'
-              +' {:}, {:}, this may be unstable'.format(times.size, 2*order+1))
-    #Adopt a default reference Julian Date, if one wasn't explicitly given
-    if jd0 == None: jd0 = np.min(times)
-    time0 = times-jd0
-
-    #Setup a grid of periods to search
-    periods = np.logspace(np.log10(pStart), np.log10(pEnd), pNum)
-
-    #We want to fit the data using sinusoid/Fourier series with fixed period,
-    #and run through a grid of periods.  The function is then
-    # y = a0 + sum_n( a_n sin(n*2*pi*T/P) + b_n cos(n*2*pi*T/P) )
-    #This can be turned into a linear problem like M.x = y (. == dot product)
-    #by using x = [a0, a1, b1, a2, b2, ... a_npar, b_npar]
-    #Then M has dimension (num observed times, free Fourier parameters) and
-    # M[T,:] has rows [1, sin(1*2*pi*T/P), cos(1*2*pi*T/P), 
-    #                     sin(2*2*pi*T/P), cos(2*2*pi*T/P),
-    #                     sin(3*2*pi*T/P), cos(3*2*pi*T/P) ...,
-    #                     sin(npar*2*pi*T/P), cos(npar*2*pi*T/P]
-
-    ##This is the relatively clear but slow version,
-    ##which loops over periods to fill in the chi^2 array for the periodogram
[…]
