+++ harmonicFit/hfNotes.txt
@@ -1,3 +1,0 @@
-WARNING: This branch is for code in development!!
-
-Notes for new harmonicFit branch.

[…]
