### 2025-06-27 ab7e4bb
Update the docstring of LSD.norm function
To account for passing an array of len(lsd).
