### 2025-06-27 5d855b2
Capitalize lsd as per Colin's suggestion
