### 2025-06-27 7551aaf
adding the harmonic fitting folder to this branch
