### 2025-07-02 #7891 @warrickball
Great, thanks!

So, first thing, congratulations @folsomcp: your submission has been accepted for publication in JOSS! I'll start start working through the post-review process in the next few days. I'll generate a checklist now with things for both of us to do as we work towards publication.

The main thing is to create a tagged release of the software and upload it to a suitable archive service (usually Zenodo). The Zenodo archive must have the same title and should have the same authors as the paper. Then let me know the tagged version number and archive DOI here so I can update the submission metadata.
