### 2025-07-06 4a8fc64
Editorial tweaks for JOSS paper
