+++ paper/paper.md
@@ -1,19 +1,21 @@
 ---
 title: 'SpecpolFlow: a new software package for spectropolarimetry using Python'
 tags:
-  - Spectropolarimetry
-  - Stellar magnetic fields
-  - Astronomy software
-  - Open source software
-  - Astronomy data visualization
+  - spectropolarimetry
+  - stellar magnetic fields
+  - astronomy software
+  - open source software
+  - astronomy data visualization
 authors:
-  - name: Colin P. Folsom^[Co-first author]
+  - name: Colin P. Folsom
     orcid: 0000-0002-9023-7890
     corresponding: true
     affiliation: 1
-  - name: Christiana Erba^[Co-first author]
+    equal-contrib: true
+  - name: Christiana Erba
     orcid: 0000-0003-1299-8878
     affiliation: 2
+    equal-contrib: true
   - name: Veronique Petit
     orcid: 0000-0002-5633-7548
     affiliation: 3
@@ -22,7 +24,7 @@ authors:
     affiliation: 3
   - name: Patrick Stanley
     orcid: 0000-0002-0378-0140
-    affiliation: 3 
+    affiliation: 3
   - name: Tali Natan
     orcid: 0000-0002-7703-6701
     affiliation: 3
@@ -40,8 +42,8 @@ authors:
     affiliation: 3
   - name: Marisol Catalan Olais
     orcid: 0009-0006-2442-6235
-    affiliation: 3    
-   
+    affiliation: 3
+
 affiliations:
  - name: Tartu Observatory, University of Tartu, Observatooriumi 1, 61602, Toravere, Estonia
    index: 1
@@ -55,7 +57,7 @@ affiliations:
    index: 5
  - name: LESIA, Observatoire de Paris, PSL University, CNRS, Sorbonne Université, Université Paris Cité, 5 place Jules Janssen, 92195 Meudon, France
    index: 6
- 
+
 date: 25 October 2024
 bibliography: paper.bib
 
@@ -63,19 +65,19 @@ bibliography: paper.bib
 
 # Summary
 
-Spectropolarimetry, the observation of polarization and intensity as a function of wavelength, is a powerful tool in stellar astrophysics. It is particularly useful for characterizing stars and circumstellar material, and for tracing the influence of magnetic fields on a host star and its environment. Maintaining modern, flexible, and accessible computational tools that enable spectropolarimetric studies is thus essential. The `SpecpolFlow` package is a new, completely Pythonic workflow for analyzing stellar spectropolarimetric observations. Its suite of tools provides a user-friendly interface for working with data from an assortment of instruments and telescopes. `SpecpolFlow` contains tools for spectral normalization and visualization, the extraction of Least-Squares Deconvolution (LSD) profiles, the generation and optimization of line masks for LSD analyses, and the calculation of longitudinal magnetic field measurements from the LSD profiles. It also provides Python classes for the manipulation of spectropolarimetric products. The `SpecpolFlow` website includes an array of tutorials that guide users through common analysis procedures using the software. `SpecpolFlow` is distributed as a free, open-source package, with fully documented tools (via an API and command line interface) which are actively maintained by a team of contributors. 
[…]
