+++ paper/paper.bib
@@ -58,7 +58,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{Reiners2012,
        author = {{Reiners}, Ansgar},
-        title = "{Observations of Cool-Star Magnetic Fields}",
+        title = {Observations of cool-star magnetic fields},
       journal = {Living Reviews in Solar Physics},
      keywords = {stars: activity, magnetic fields, Spectral Type, Zeeman Splitting, Magnetic Field Measurement, Stellar Surface, Rossby Number, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2012,
@@ -77,7 +77,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{DonatiLandstreet2009,
        author = {{Donati}, J. -F. and {Landstreet}, J.~D.},
-        title = "{Magnetic Fields of Nondegenerate Stars}",
+        title = {Magnetic fields of nondegenerate stars},
       journal = {Annual Review of Astronomy & Astrophysics},
      keywords = {Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2009,
@@ -164,7 +164,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{Ferrario2015,
        author = {{Ferrario}, Lilia and {de Martino}, Domitilla and {G{\"a}nsicke}, Boris T.},
-        title = "{Magnetic White Dwarfs}",
+        title = {Magnetic white dwarfs},
       journal = {Space Science Reviews},
      keywords = {Magnetic fields, Magnetic white dwarfs, Magnetic cataclysmic variables, Binary systems, Astrophysics - Solar and Stellar Astrophysics},
          year = 2015,
@@ -328,3 +328,25 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Harris2020,
+ title         = {Array programming with {NumPy}},
+ author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
+                 van der Walt and Ralf Gommers and Pauli Virtanen and David
+                 Cournapeau and Eric Wieser and Julian Taylor and Sebastian
+                 Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
+                 and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
+                 Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
+                 R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
+                 G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
+                 Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
+                 Travis E. Oliphant},
+ year          = {2020},
+ month         = sep,
+ journal       = {Nature},
+ volume        = {585},
+ number        = {7825},
+ pages         = {357--362},
+ doi           = {10.1038/s41586-020-2649-2},
+ publisher     = {Springer Science and Business Media {LLC}},
+ url           = {https://doi.org/10.1038/s41586-020-2649-2}
+}
