### 2025-07-08 ecad0ea
Changed the formatting of SpecpolFlow in the paper to use a regular font, following Warrick Ball's suggestion
