+++ paper/paper.md
@@ -65,7 +65,7 @@ bibliography: paper.bib
 
 # Summary
 
-Spectropolarimetry, the observation of polarization and intensity as a function of wavelength, is a powerful tool in stellar astrophysics. It is particularly useful for characterizing stars and circumstellar material, and for tracing the influence of magnetic fields on a host star and its environment. Maintaining modern, flexible, and accessible computational tools that enable spectropolarimetric studies is thus essential. The `SpecpolFlow` package is a new, completely Pythonic workflow for analyzing stellar spectropolarimetric observations. Its suite of tools provides a user-friendly interface for working with data from an assortment of instruments and telescopes. `SpecpolFlow` contains tools for spectral normalization and visualization, the extraction of Least-Squares Deconvolution (LSD) profiles, the generation and optimization of line masks for LSD analyses, and the calculation of longitudinal magnetic field measurements from the LSD profiles. It also provides Python classes for the manipulation of spectropolarimetric products. The `SpecpolFlow` website includes an array of tutorials that guide users through common analysis procedures using the software. `SpecpolFlow` is distributed as a free, open-source package, with fully documented tools (via an API and command line interface) which are actively maintained by a team of contributors.
+Spectropolarimetry, the observation of polarization and intensity as a function of wavelength, is a powerful tool in stellar astrophysics. It is particularly useful for characterizing stars and circumstellar material, and for tracing the influence of magnetic fields on a host star and its environment. Maintaining modern, flexible, and accessible computational tools that enable spectropolarimetric studies is thus essential. The SpecpolFlow package is a new, completely Pythonic workflow for analyzing stellar spectropolarimetric observations. Its suite of tools provides a user-friendly interface for working with data from an assortment of instruments and telescopes. SpecpolFlow contains tools for spectral normalization and visualization, the extraction of Least-Squares Deconvolution (LSD) profiles, the generation and optimization of line masks for LSD analyses, and the calculation of longitudinal magnetic field measurements from the LSD profiles. It also provides Python classes for the manipulation of spectropolarimetric products. The SpecpolFlow website includes an array of tutorials that guide users through common analysis procedures using the software. SpecpolFlow is distributed as a free, open-source package, with fully documented tools (via an API and command line interface) which are actively maintained by a team of contributors.
 
 # Statement of need
 
@@ -77,9 +77,9 @@ Several successful programs supporting spectropolarimetric analyses exist in the
 
 # Overview of SpecpolFlow
 
-The `SpecpolFlow` package is a modernized, unified revitalization of the tools that have preceded it. The software is open source, well documented, and the code itself is extensively commented and designed to be readable. `SpecpolFlow` produces results that are consistent with previous proprietary codes implementing similar algorithms. It produces LSD profiles that are consistent with the code of @Donati1997 and `iLSD`, and it produces $\langle B_z \rangle$ values that are consistent with the code of @Wade2000.
+The SpecpolFlow package is a modernized, unified revitalization of the tools that have preceded it. The software is open source, well documented, and the code itself is extensively commented and designed to be readable. SpecpolFlow produces results that are consistent with previous proprietary codes implementing similar algorithms. It produces LSD profiles that are consistent with the code of @Donati1997 and `iLSD`, and it produces $\langle B_z \rangle$ values that are consistent with the code of @Wade2000.
 
-`SpecpolFlow` provides a toolkit with an ensemble of Python functions that:
+SpecpolFlow provides a toolkit with an ensemble of Python functions that:
 
[…]
