### 2025-07-08 af3a30f
setting the version number to 1.0 in advance of the full release
