### 2025-07-11 #7891 @folsomcp
Thanks to the reviewers for their help getting us to the acceptance stage!

We've accepted the first round of editorial tweaks to the paper.  @warrickball is now an appropriate time to create a tagged release and upload it to Zenodo?  (Basically just double checking that I'm not supposed to wait for any further editorial comments).
