### 2025-07-11 #7891 @warrickball
Sorry for confusing things: you can go ahead with the tagged release quite independently of the editorial tweaks. We would only want to revise the Zenodo archive if the code itself changes; we don't mind the paper is slightly out of sync. Normally the Zenodo archive is created clearly before final editorial stuff but because I'm editor *and* associate EiC I've muddied the water a bit.
