### 2025-07-13 8e76cbb
switching to PyPI's new license meta-data tags
