### 2025-07-13 64e46f2
added a mention of our paper and the pre-print to the README
