### 2025-07-14 #7891 @folsomcp
I've created a tagged release for the SpecpolFlow as version 1.0.0, tagged with v1.0.0,  https://github.com/folsomcp/specpolFlow/releases/tag/v1.0.0

And I've archived a copy of this release on Zenodo, https://zenodo.org/records/15875240  with the DOI 10.5281/zenodo.15875240  The title and authors in Zenodo should match the paper.

So I think we are ready to proceed @warrickball but if I missed anything please let me know.
