### 2025-07-19 #7891 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
